+++ code/libraries/utils.h
@@ -10,13 +10,13 @@ using DMatrix = vector<vector<double>>;
 
 namespace utils
 {
-    IntMatrix getRand(int n, int minVal, int maxVal)
+    IntMatrix getRand(unsigned long int n, int minVal, int maxVal)
     // Generates an n x n matrix with random integers between minVal and maxVal
     {
         IntMatrix grid(n, vector<int>(n));
-        for (int i = 0; i < n; i++)
+        for (unsigned long int i = 0; i < n; i++)
         {
-            for (int j = 0; j < n; j++)
+            for (unsigned long int j = 0; j < n; j++)
             {
                 grid[i][j] = rand() % (maxVal - minVal + 1) + minVal;
             }
@@ -29,9 +29,9 @@ namespace utils
     {
         int n = intMatrix.size();
         DMatrix doubleMatrix(n, vector<double>(n));
-        for (int i = 0; i < n; i++)
+        for (unsigned long int i = 0; i < n; i++)
         {
-            for (int j = 0; j < n; j++)
+            for (unsigned long int j = 0; j < n; j++)
             {
                 doubleMatrix[i][j] = static_cast<double>(intMatrix[i][j]);
             }
